jCUTE can be built with vagrant up command.
jCUTE can now be built in a 64-bit ubuntu linux container using the Vagrant. Run 'vagrant up' to build jCUTE and 'vagrant ssh' to access the machine. This repository is locally synced with the machine's folder /vagrant, so any changes to the /vagrant folder will be felt in the host.